+++ source/main/ipc/getHardwareWalletChannel.js
@@ -682,7 +682,7 @@ export const handleHardwareWalletRequests = async (
     }
   });
 
-  // @TODO - validityIntervalStartStr is not working with Cardano App 2.1.0
+  // @TODO - validityIntervalStart is not working with Cardano App 2.1.0
   signTransactionLedgerChannel.onRequest(async (params) => {
     const {
       inputs,
@@ -695,7 +695,6 @@ export const handleHardwareWalletRequests = async (
       withdrawals,
       auxiliaryData,
       devicePath,
-      // validityIntervalStartStr,
       signingMode,
     } = params;
     logger.info('[HW-DEBUG] SIGN Ledger transaction');
@@ -721,7 +720,6 @@ export const handleHardwareWalletRequests = async (
           certificates,
           withdrawals,
           auxiliaryData,
-          // validityIntervalStart,
         },
       });
       return Promise.resolve(signedTransaction);