+++ source/renderer/app/api/api.js
@@ -160,6 +160,7 @@ import type {
   CreateExternalTransactionResponse,
   GetWithdrawalsRequest,
   GetWithdrawalsResponse,
+  VotingMetadataType,
   ICOPublicKeyParams,
 } from './transactions/types';
 
@@ -1044,6 +1045,7 @@ export default class AdaApi {
     rewardsBalance: BigNumber,
     payments?: CoinSelectionsPaymentRequestType,
     delegation?: CoinSelectionsDelegationRequestType,
+    metadata?: VotingMetadataType,
   }): Promise<CoinSelectionsResponse> => {
     logger.debug('AdaApi::selectCoins called', {
       parameters: filterLogData(request),
@@ -1055,6 +1057,7 @@ export default class AdaApi {
       walletBalance,
       availableBalance,
       rewardsBalance,
+      metadata,
     } = request;
     try {
       let data;
@@ -1078,6 +1081,7 @@ export default class AdaApi {
             },
           ],
           withdrawal: TransactionWithdrawal,
+          metadata: metadata || null,
         };
       } else {
         throw new Error('Missing parameters!');
@@ -1186,6 +1190,7 @@ export default class AdaApi {
         fee: fee.dividedBy(LOVELACES_PER_ADA),
         deposits: deposits.dividedBy(LOVELACES_PER_ADA),
         depositsReclaimed: depositsReclaimed.dividedBy(LOVELACES_PER_ADA),
+        metadata: response.metadata || null,
       };
 
       logger.debug('AdaApi::selectCoins success', { extendedResponse });